% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launch}
\alias{launch}
\title{Launch visualization}
\usage{
launch(prompt = interactive())
}
\arguments{
\item{prompt}{whether or not to prompt the default browser to open the app. 
By default, a browser is prompted when R is being used interactively.}
}
\description{
Launch visualization
}
\examples{

launch()
}
\author{
Carson Sievert
}
